+++ src/mitim_tools/gs_tools/utils/GEQplotting.py
@@ -7,8 +7,8 @@ from IPython import embed
 
 #TODO: add current profiles and flux-surface average fields to megpy and restore plots
 
-def compareGeqdsk(geqdsks, fn=None, extraLabel="", plotAll=True, labelsGs=None):
-    
+def compareGeqdsk(geqdsks, fn=None, extraLabel="", plotAll=True, labelsGs=None, tab_color=None):
+
     if fn is None:
         from mitim_tools.misc_tools.GUItools import FigureNotebook
         fn = FigureNotebook("GEQDSK Notebook", geometry="1600x1000")
@@ -23,15 +23,15 @@ def compareGeqdsk(geqdsks, fn=None, extraLabel="", plotAll=True, labelsGs=None):
     # -----------------------------------------------------------------------------
     if plotAll:
         for i, g in enumerate(geqdsks):
-            _ = g.plot(fn=fn, extraLabel=f"{labelsGs[i]} - " if len(labelsGs) > i else "")
+            _ = g.plot(fn=fn, extraLabel=f"{labelsGs[i]} - " if len(labelsGs) > i else "", tab_color=tab_color)
 
     if len(geqdsks) < 2:
         return None, fn
-    
+
     # -----------------------------------------------------------------------------
     # Compare in same plot - Surfaces
     # -----------------------------------------------------------------------------
-    fig = fn.add_figure(label=extraLabel + "Comp. - Surfaces")
+    fig = fn.add_figure(label=extraLabel + "Comp. - Surfaces", tab_color=tab_color)
 
     grid = plt.GridSpec(2, 3, hspace=0.3, wspace=0.3)
     ax1 = fig.add_subplot(grid[:, 0])
@@ -50,7 +50,7 @@ def compareGeqdsk(geqdsks, fn=None, extraLabel="", plotAll=True, labelsGs=None):
     # -----------------------------------------------------------------------------
     # Compare in same plot - Surfaces
     # -----------------------------------------------------------------------------
-    fig = fn.add_figure(label=extraLabel + "Comp. - Plasma")
+    fig = fn.add_figure(label=extraLabel + "Comp. - Plasma", tab_color=tab_color)
     grid = plt.GridSpec(2, 4, hspace=0.3, wspace=0.3)
 
     ax_plasma = [